refactor: move union_find family from tree/ to data_structure/
Union-Find (DSU) is a disjoint-set data structure, not a tree algorithm — its
rooted-forest internals are an implementation detail. Move the whole family so
it lives with related data structures (e.g. offline_dynamic_connectivity):

  lib/tree/{union_find,weighted_union_find,undo_union_find,dynamic_union_find}.hpp
  -> lib/data_structure/

dsu_on_tree.hpp stays in tree/: it is a tree query technique over Graph<T>,
not a Union-Find data structure.

All 17 include sites updated (lib/graph/kruskal.hpp,
lib/data_structure/offline_dynamic_connectivity.hpp, and tests). The four
headers are self-contained, so no other paths change. Verified all dependents
compile with g++ 13.3 -std=gnu++23.
