+++ test/aoj/challenge/2559.2.test.cpp
@@ -9,7 +9,7 @@
 #include <vector>
 #include "graph/graph.hpp"
 #include "heap/leftist_heap.hpp"
-#include "tree/union_find.hpp"
+#include "data_structure/union_find.hpp"
 
 int main(void) {
     int n, m;
