+++ test/aoj/jag/dsu_on_tree.test.cpp
@@ -5,7 +5,7 @@
 #include <utility>
 #include <vector>
 #include "graph/graph.hpp"
-#include "tree/union_find.hpp"
+#include "data_structure/union_find.hpp"
 
 int main(void) {
     int n, k;
