+++ test/aoj/jag/dynamic_union_find.test.cpp
@@ -1,5 +1,5 @@
 // competitive-verifier: PROBLEM https://onlinejudge.u-aizu.ac.jp/problems/2995
-#include "tree/dynamic_union_find.hpp"
+#include "data_structure/dynamic_union_find.hpp"
 #include <iostream>
 #include <unordered_map>
 #include <utility>
