+++ test/aoj/joi/undo_union_find.test.cpp
@@ -1,5 +1,5 @@
 // competitive-verifier: PROBLEM https://onlinejudge.u-aizu.ac.jp/problems/0723
-#include "tree/undo_union_find.hpp"
+#include "data_structure/undo_union_find.hpp"
 #include <algorithm>
 #include <iostream>
 #include <map>
