+++ test/yosupo/data_structure/undo_union_find.test.cpp
@@ -1,5 +1,5 @@
 // competitive-verifier: PROBLEM https://judge.yosupo.jp/problem/persistent_unionfind
-#include "tree/undo_union_find.hpp"
+#include "data_structure/undo_union_find.hpp"
 #include <iostream>
 #include <utility>
 #include <vector>
