+++ test/yukicoder/0583.test.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <set>
 #include "data_structure/flip_set.hpp"
-#include "tree/union_find.hpp"
+#include "data_structure/union_find.hpp"
 
 int main(void) {
     int n, m;
